backend/requirements.txt
fastapi
openai
requests
uvicorn[standard]
gunicorn
python-multipart
azure-cosmos
